tutorials/ner_pytorch_medical/requirements.txt
spacy-streamlit>=1.0.0a0
spacy-transformers
torch==1.8.1
streamlit
presidio-analyzer==2.2.1
presidio-anonymizer==2.2.1
